move generate data util from vuu-data to showcase, add faker as depdency of showcase
