+++ vuu-ui/showcase/package.json
@@ -16,6 +16,7 @@
   "author": "heswell",
   "license": "Apache-2.0",
   "dependencies": {
+    "@faker-js/faker": "^8.0.2",
     "@finos/vuu-data-ag-grid": "0.0.26",
     "@finos/vuu-datagrid": "0.0.26",
     "@finos/vuu-filters": "0.0.26",
@@ -23,9 +24,9 @@
     "@finos/vuu-theme": "0.0.26",
     "@finos/vuu-utils": "0.0.26",
     "@salt-ds/core": "1.8.0",
+    "@salt-ds/icons": "1.5.1",
     "@salt-ds/lab": "1.0.0-alpha.15",
     "@salt-ds/theme": "1.7.1",
-    "@salt-ds/icons": "1.5.1",
     "classnames": "^2.3.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
